improve_initial_state
