+++ examples/main.py
@@ -14,10 +14,4 @@ if __name__ == "__main__":
 
     plot_basis_states(system, config)
     plot_state_against_t(system, config, n=1000, step=500)
-    # plot_kernel(system, config, temperature=0.0001)
-    # plot_lindblad_operator(system, config)
-    # plot_thermal_occupation(system, config)
-    # plot_system_eigenstates(system, config)
-    # plot_stochastic_occupation(system, config, n=1000, step=500)
     plot_initial_state(system, config)
-    # plot_noise_operator(system, config, temperature=150.0)
