+++ reduced_state_caldeira_leggett/plot.py
@@ -220,11 +220,6 @@ def plot_stochastic_occupation(
     )
     hamiltonian = get_hamiltonian(system, config)
 
-    # fig0, ax0 = plot_all_band_occupations(hamiltonian, states)
-
-    # fig1, ax1 = fig0, ax0
-    # fig1, ax1, _ani = animate_all_band_occupations(hamiltonian, states)
-
     fig2, ax2, line = plot_average_band_occupation(hamiltonian, states)
 
     for ax in [ax2]:
@@ -234,8 +229,6 @@ def plot_stochastic_occupation(
 
         ax.legend([line], ["Boltzmann occupation"])
 
-    # fig0.show()
-    # fig1.show()
     fig2.show()
     input()
